modules/MT5LiveDataClient/requirements.txt
python-socketio[client]==5.11.2
psycopg2-binary==2.9.9